Updated LeagueDetailPage.tsx
